+++ packages/example/src/components/connector/connectors-setup.ts
@@ -1,24 +1,19 @@
 import { NetworkType as TezosNetwork } from "@airgap/beacon-sdk"
-import { EthereumWallet, FlowWallet, TezosWallet } from "@rarible/sdk-wallet"
 import { RaribleSdkEnvironment } from "@rarible/sdk/build/config/domain"
-import Web3 from "web3"
-import { Web3Ethereum } from "@rarible/web3-ethereum"
 import {
-	AbstractConnectionProvider,
 	ConnectionProvider,
 	Connector,
-	EthereumProviderConnectionResult,
 	IConnectorStateProvider,
 	InjectedWeb3ConnectionProvider,
 } from "@rarible/connector"
-import { FclConnectionProvider, FlowProviderConnectionResult } from "@rarible/connector-fcl"
+import { FclConnectionProvider } from "@rarible/connector-fcl"
 import { MEWConnectionProvider } from "@rarible/connector-mew"
-import { BeaconConnectionProvider, TezosProviderConnectionResult } from "@rarible/connector-beacon"
+import { BeaconConnectionProvider } from "@rarible/connector-beacon"
 import { TorusConnectionProvider } from "@rarible/connector-torus"
 import { WalletLinkConnectionProvider } from "@rarible/connector-walletlink"
 import { WalletConnectConnectionProvider } from "@rarible/connector-walletconnect"
-import type { IWalletAndAddress } from "./wallet-connetion"
-import { Blockchain } from "@rarible/api-client"
+import type { IWalletAndAddress } from "@rarible/connector-helper"
+import { mapEthereumWallet, mapFlowWallet, mapTezosWallet } from "@rarible/connector-helper";
 // import { FortmaticConnectionProvider } from "@rarible/connector-fortmatic"
 // import { PortisConnectionProvider } from "@rarible/connector-portis"
 
@@ -91,47 +86,6 @@ function environmentToTezosNetwork(environment: RaribleSdkEnvironment) {
 				network: TezosNetwork.HANGZHOUNET
 			}
 	}
-}
-
-function mapEthereumWallet<O>(provider: AbstractConnectionProvider<O, EthereumProviderConnectionResult>): ConnectionProvider<O, IWalletAndAddress> {
-	return provider.map(state => ({
-		wallet: new EthereumWallet(new Web3Ethereum({ web3: new Web3(state.provider), from: state.address })),
-		address: state.address,
-		blockchain: getEvmBlockchain(state.chainId)
-	}))
-}
-
-function getEvmBlockchain(chainId: number): Blockchain.POLYGON | Blockchain.ETHEREUM {
-	switch (chainId) {
-		case 137: return Blockchain.POLYGON
-		case 80001: return Blockchain.POLYGON
-		case 300501: return Blockchain.POLYGON
-		case 200501: return Blockchain.POLYGON
-		default: return Blockchain.ETHEREUM
-	}
-}
-
-function mapFlowWallet<O>(provider: AbstractConnectionProvider<O, FlowProviderConnectionResult>): ConnectionProvider<O, IWalletAndAddress> {
-	return provider.map(state => ({
-		wallet: new FlowWallet(state.fcl),
-		address: state.address,
-		blockchain: Blockchain.FLOW,
-	}))
-}
-
-function mapTezosWallet<O>(provider: AbstractConnectionProvider<O, TezosProviderConnectionResult>): ConnectionProvider<O, IWalletAndAddress> {
-	return provider.map(async state => {
-		const {
-			beacon_provider: createBeaconProvider
-		} = await import("@rarible/tezos-sdk/dist/providers/beacon/beacon_provider")
-		const provider = await createBeaconProvider(state.wallet as any, state.toolkit)
-
-		return {
-			wallet: new TezosWallet(provider),
-			address: state.address,
-			blockchain: Blockchain.TEZOS
-		}
-	})
 }
 
 const state: IConnectorStateProvider = {