+++ packages/example/src/components/connector/wallet-connetion.ts
@@ -1,8 +1,0 @@
-import { BlockchainWallet } from "@rarible/sdk-wallet"
-import { Blockchain } from "@rarible/api-client"
-
-export interface IWalletAndAddress {
-	wallet: BlockchainWallet
-	address: string
-	blockchain: Blockchain
-}